requirements.txt
ast2json==0.3.0
libcst==1.1.0
sentence-transformers==2.2.2
chromadb==0.4.22
langchain==0.1.0
torch==2.5.1
google-generativeai==0.3.2
python-dotenv==1.0.0